move modules genre to a new folder
