+++ cmd/main.go
@@ -5,6 +5,7 @@ import (
 	"movie-app-go/database"
 	// "movie-app-go/database/seed"
 	"movie-app-go/internal/modules/iam"
+	"movie-app-go/internal/modules/genre"
 	"movie-app-go/internal/modules/studio"
 	"movie-app-go/internal/modules/movie"
 	"os"
@@ -36,12 +37,14 @@ func main() {
 	iamModule := iam.NewIAMModule(db)
 	studioModule := studio.NewStudioModule(db)
 	movieModule := movie.NewMovieModule(db)
+	genreModule := genre.NewGenreModule(db)
 
 	// Setup Gin
 	r := gin.Default()
 	iam.RegisterRoutes(r.Group("/api"), iamModule)
 	studio.RegisterRoutes(r.Group("/api"), studioModule)
 	movie.RegisterRoutes(r.Group("/api"), movieModule)
+	genre.RegisterRoutes(r.Group("/api"), genreModule)
 
 	// Run server
 	r.Run(":" + port)
